Move StorageManifestDao to the bag_tracker app
