+++ bag_tagger/src/test/scala/uk/ac/wellcome/platform/storage/bag_tagger/fixtures/BagTaggerFixtures.scala
@@ -11,13 +11,13 @@ import uk.ac.wellcome.platform.archive.bag_tracker.fixtures.{
   BagTrackerFixtures,
   StorageManifestDaoFixture
 }
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.StorageManifestDao
 import uk.ac.wellcome.platform.archive.common.BagRegistrationNotification
 import uk.ac.wellcome.platform.archive.common.fixtures.OperationFixtures
 import uk.ac.wellcome.platform.archive.common.storage.models.{
   StorageManifest,
   StorageManifestFile
 }
-import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestDao
 import uk.ac.wellcome.platform.storage.bag_tagger.services.{
   ApplyTags,
   BagTaggerWorker,
