+++ bag_tracker/docker-compose.yml
@@ -1,0 +1,10 @@
+dynamodb:
+  image: peopleperhour/dynamodb
+  ports:
+    - "45678:8000"
+s3:
+  image: "zenko/cloudserver:8.1.8"
+  environment:
+    - "S3BACKEND=mem"
+  ports:
+    - "33333:8000"
