+++ bag_tracker/src/main/scala/uk/ac/wellcome/platform/archive/bag_tracker/config/builders/StorageManifestDaoBuilder.scala
@@ -3,8 +3,8 @@ package uk.ac.wellcome.platform.archive.bag_tracker.config.builders
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.amazonaws.services.s3.AmazonS3
 import com.typesafe.config.Config
-import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestDao
-import uk.ac.wellcome.platform.archive.common.storage.services.dynamo.DynamoStorageManifestDao
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.StorageManifestDao
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.dynamo.DynamoStorageManifestDao
 import uk.ac.wellcome.storage.typesafe.{DynamoBuilder, S3Builder}
 
 object StorageManifestDaoBuilder {
