+++ bag_tracker/src/main/scala/uk/ac/wellcome/platform/archive/bag_tracker/services/CreateBag.scala
@@ -4,8 +4,8 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives.complete
 import akka.http.scaladsl.server.Route
 import grizzled.slf4j.Logging
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.StorageManifestDao
 import uk.ac.wellcome.platform.archive.common.storage.models.StorageManifest
-import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestDao
 
 trait CreateBag extends Logging {
   val storageManifestDao: StorageManifestDao
