+++ bag_tracker/src/main/scala/uk/ac/wellcome/platform/archive/bag_tracker/services/GetLatestBag.scala
@@ -6,8 +6,8 @@ import akka.http.scaladsl.server.Route
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.json.JsonUtil._
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.StorageManifestDao
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
-import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestDao
 import uk.ac.wellcome.storage.NoVersionExistsError
 
 trait GetLatestBag extends Logging {
