+++ bag_tracker/src/main/scala/uk/ac/wellcome/platform/archive/bag_tracker/services/LookupBagVersions.scala
@@ -10,8 +10,8 @@ import uk.ac.wellcome.platform.archive.bag_tracker.models.{
   BagVersionEntry,
   BagVersionList
 }
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.StorageManifestDao
 import uk.ac.wellcome.platform.archive.common.bagit.models.{BagId, BagVersion}
-import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestDao
 
 import scala.concurrent.ExecutionContext
