+++ bag_tracker/src/test/scala/uk/ac/wellcome/platform/archive/bag_tracker/storage/memory/MemoryStorageManifestDaoTest.scala
@@ -1,12 +1,12 @@
-package uk.ac.wellcome.platform.archive.common.storage.services.memory
+package uk.ac.wellcome.platform.archive.bag_tracker.storage.memory
 
 import uk.ac.wellcome.fixtures.TestWith
-import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
-import uk.ac.wellcome.platform.archive.common.storage.models.StorageManifest
-import uk.ac.wellcome.platform.archive.common.storage.services.{
+import uk.ac.wellcome.platform.archive.bag_tracker.storage.{
   StorageManifestDao,
   StorageManifestDaoTestCases
 }
+import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
+import uk.ac.wellcome.platform.archive.common.storage.models.StorageManifest
 import uk.ac.wellcome.storage.store.memory.MemoryVersionedStore
 
 class MemoryStorageManifestDaoTest
