+++ common/docker-compose.yml
@@ -4,10 +4,6 @@ s3:
     - "S3BACKEND=mem"
   ports:
     - "33333:8000"
-dynamodb:
-  image: peopleperhour/dynamodb
-  ports:
-    - "45678:8000"
 azurite:
   image: "mcr.microsoft.com/azure-storage/azurite"
   ports:
